+++ R/layers-core.R
@@ -502,6 +502,8 @@ compose_layer.keras.models.Sequential <- function(object, layer) {
   object
 }
 
+compose_layer.keras.engine.sequential.Sequential <- compose_layer.keras.models.Sequential
+
 compose_layer.python.builtin.object <- function(object, layer) {
   if (is.function(layer))
     layer(object)